refactor: extract shared tab close dialogs into TabCloseHelper
Move duplicated tab close confirmation logic from PaneLeafView and
ContentView+Helpers into a shared TabCloseHelper enum. Both call sites
now delegate to the helper, eliminating ~80 lines of copy-paste code.
